backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI / Anthropic
anthropic>=0.18.1

# Storage (S3 compatible)
boto3>=1.34.0

# HTTP client
httpx>=0.26.0

# Utils
python-dateutil>=2.8.2
